requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

python-dotenv==1.0.1
requests==2.32.3

supabase==2.6.0

line-bot-sdk==3.11.0

gspread==6.1.4
google-auth==2.33.0

sentence-transformers==3.0.1
torch>=2.2.0
numpy==1.26.4
python-dateutil==2.9.0.post0